Avoid some warnings - 'array subscript has type char'

+++ libretro-common/net/net_http.c
@@ -226,7 +226,7 @@ bool net_http_connection_done(struct http_connection_t *conn)
    if (*conn->scan == ':')
    {
 
-      if (!isdigit(conn->scan[1]))
+      if (!isdigit((int)conn->scan[1]))
          return false;
 
       conn->port = strtoul(conn->scan + 1, &conn->scan, 10);
